requirements.txt
tensorflow
numpy
matplotlib
pyyaml
mlflow
pytest